Get rid of all uses of componentWillUpdate, componentWillMount.
Factor out redundant error handling into a single helper function. This function also handles all the cases for axios errors.

Delete EditDeckModal since the SettingsTab replaced it.

Change the tests to runInBand, fixing the slow tests.

For the ResetPassword page, redirect if the user is already authenticated or there is no token. Also, remove the token from state since props
is sufficient.

Factor out analogous HOC to ReqAuth: AlreadyAuth. Redirects if the user is logged in.

Convert the Settings component into a simpler functional component.

Eject the axios interceptors on dismount for the withErrors HOC.